fixed the dependency cycle
